+++ docs/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Gopjrt Changelog
 
-# Next
+# v0.8.1 = 2025/09/21 Changes in support [github.com/openxla/stablehlo](https://github.com/openxla/stablehlo) 
 
 * Moved `stablehlo` to its own repository.
 * Updated dependencies.
+* Package `dtypes`:
+    * Added `DType.Bits` field indicating the number of bits in the data type -- supporting sub-byte types including
+      quantized types.
+    * Added `DType.IsPromotableTo(target)` to check if types are promotable to the target type -- it follows the StableHLO rule.
+    * Cleaned up generators for `dtypes` package.
+* Package `xlabuilder`:
+  * Deprecated `SelectAndScatterSum`.
 
 # v0.8.0 - 2025/08/20 Convolution API changes